move several components to `common` folder to structure repo better - part of #325
