+++ frontend/src/components/common/index.js
@@ -1,0 +1,6 @@
+import ActionButton from './actionButton'
+import ActionLink from './actionLink'
+import ChangeStatus from './changeStatus'
+import ErrorMessage from './errorMessage'
+
+export { ActionButton, ActionLink, ChangeStatus, ErrorMessage }
