+++ frontend/src/pages/_app.js
@@ -2,9 +2,9 @@ import React from 'react'
 import App, { Container } from 'next/app'
 import { Provider } from 'react-redux'
 import withRedux from 'next-redux-wrapper'
-import configureStore from '../store/appStore'
-import ErrorMessage from '../components/errorMessage'
 import { PageTransition } from 'next-page-transitions'
+import configureStore from '../store/appStore'
+import { ErrorMessage } from '../components/common'
 
 export default withRedux(configureStore, {
   debug: process.env.DEBUG_REDUX === undefined ? false : process.env.DEBUG_REDUX === 'true'
