+++ frontend/src/pages/dashboard.js
@@ -21,8 +21,7 @@ import Nav from '../components/nav'
 import CandidateStatus from '../components/candidateStatus'
 import CandidateLinksBadge from '../components/candidateLinksBadge'
 import FilterComponent from '../components/filterComponent'
-import ChangeStatus from '../components/changeStatus'
-import ErrorMessage from '../components/errorMessage'
+import { ChangeStatus, ErrorMessage } from '../components/common'
 
 import { getCandidates, setCandidateStatus } from '../utils/api'
 import { avgInterviewScore, compareByAvgInterviewScore, getNumOfInterviews } from '../utils/core'
