+++ frontend/src/pages/interview.js
@@ -20,7 +20,7 @@ import { bindActionCreators } from 'redux'
 import Router from 'next/router'
 import { fetchAllCandidates, addFilter, removeFilter, fetchCandidatesSuccess } from '../actions'
 import CandidateDropdown from '../components/candidateDropdown'
-import ErrorMessage from '../components/errorMessage'
+import { ErrorMessage } from '../components/common'
 import InterviewSectionCard from '../components/interviewSectionCard'
 import VerificationModal from '../components/verificationModal'
 import InterviewSectionModular from '../components/interviewSectionModular'
