+++ frontend/src/pages/interviewlist.js
@@ -9,8 +9,7 @@ import {
 } from '../utils/api'
 import CandidateInterviewsModal from '../components/candidates/candidateInterviewsModal'
 import { avgInterviewScore, interviewGetCategorySection } from '../utils/core'
-import ActionLink from '../components/actionLink'
-import ActionButton from '../components/actionButton'
+import { ActionLink, ActionButton } from '../components/common'
 import Nav from '../components/nav'
 import Head from '../components/head'
 import roundData from '../data/roundData'
