+++ frontend/src/pages/interviewportal.js
@@ -7,7 +7,7 @@ import { bindActionCreators } from 'redux'
 import { editInterview, newInterview } from '../actions'
 import { getKey, getPastInterviews, deleteInterview } from '../utils/api'
 import VerificationModal from '../components/verificationModal'
-import ActionButton from '../components/actionButton'
+import { ActionButton } from '../components/common'
 import Nav from '../components/nav'
 import Head from '../components/head'
